ml-server/requirements.txt
# FastAPI ML 서버 (ONNX Runtime — no torch/timm needed)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
onnxruntime>=1.16.0
Pillow>=10.0.0
